steps and play work without animation

+++ LayerAssignmentProject/src/src/tdojlafry/layered/layerAssignment/graphData/Position.java
@@ -2,14 +2,14 @@ package src.tdojlafry.layered.layerAssignment.graphData;
 
 public class Position{
     
-    public int x = 0;
-    public int y = 0;
+    public double x = 0;
+    public double y = 0;
     
     public Position() {
         
     }
     
-    public Position(int a, int b) {
+    public Position(double a, double b) {
         this.x = a;
         this.y = b;
     }
